+++ packages/preprocessor/test/test.config.js
@@ -5,5 +5,9 @@ module.exports = {
   },
   use: [
     "postcss-property-lookup"
-  ]
+  ],
+  autoprefixer: {
+    add: false,
+    remove: false
+  }
 };